+++ example/fireclass-express-app/src/lib/fireclass.ts
@@ -1,11 +1,5 @@
 import "reflect-metadata";
 import { createFireclass } from "@dharayush7/fireclass-js";
-import { getDb } from "./firebase";
+import { getDb } from "./firebase.js";
 
 export const { BaseModel, adapter } = createFireclass(getDb());
-
-export {
-  Collection,
-  Subcollection,
-  fireclassErrorHandler,
-} from "@dharayush7/fireclass-js";